feat: period util함수 분리
